+++ frontend/src/App.tsx
@@ -1,5 +1,5 @@
-import Navbar from "./components/Navbar";
-import ResultsView from "./components/ResultsView/ResultsView";
+import { Navbar } from "./components/Navbar";
+import { ResultsView } from "./components/Results-View";
 import { saveState } from "./redux/localStorage";
 import store from "./redux/store";