+++ frontend/src/components/shared/buttons/add-person-btn.component.tsx
@@ -1,0 +1,25 @@
+import { AddPersonIcon } from "../icons";
+
+function AddPersonBtn({
+  clickSideEffect,
+  ariaTitle,
+  className = "",
+  iconClassName = "",
+}: any) {
+  return (
+    <button
+      className={`btn ${className ? className : ""}`}
+      onClick={() => clickSideEffect()}
+      title={ariaTitle}
+    >
+      <span className="btn-wrapper">
+        <AddPersonIcon
+          className={`icons ${iconClassName ? iconClassName : ""}`}
+        />
+        <span className="btn-text">Add Person</span>
+      </span>
+    </button>
+  );
+}
+
+export { AddPersonBtn };