+++ frontend/src/components/shared/buttons/edit-btn.component.tsx
@@ -1,0 +1,23 @@
+import { PencilIcon } from "../icons";
+
+function EditBtn({
+  clickSideEffect,
+  ariaTitle,
+  className = "",
+  iconClassName = "",
+}: any) {
+  return (
+    <button
+      className={`btn ${className ? className : ""}`}
+      onClick={() => clickSideEffect()}
+      title={ariaTitle}
+    >
+      <span className="btn-wrapper">
+        <PencilIcon className={`icons ${iconClassName ? iconClassName : ""}`} />
+        <span className="btn-text">Edit</span>
+      </span>
+    </button>
+  );
+}
+
+export { EditBtn };