+++ frontend/src/components/shared/buttons/index.tsx
@@ -1,0 +1,8 @@
+export { AddBtn } from "./add-btn.component";
+export { AddPersonBtn } from "./add-person-btn.component";
+export { CloseBtn } from "./close-btn.component";
+export { DeleteBtn } from "./delete-btn.component";
+export { DeletePersonBtn } from "./delete-person-btn.component";
+export { EditBtn } from "./edit-btn.component";
+export { ResetBtn } from "./reset-btn.component";
+export { ShareBtn } from "./share-btn.component";