+++ frontend/src/hooks/did-Mount/did-mount.hook.tsx
@@ -1,0 +1,11 @@
+import { useEffect, useRef } from "react";
+
+const didMount = () => {
+  const didMount = useRef(true);
+  useEffect(() => {
+    didMount.current = false;
+  }, []);
+  return didMount.current;
+};
+
+export { didMount };